feat: optimize folder structures
