+++ index.html
@@ -5,10 +5,10 @@
     <meta charset="utf-8" />
     <title id="doc-title">Kuru Kuru~</title>
     <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1">
-    <link rel="stylesheet" href="style.css" type="text/css" />
+    <link rel="stylesheet" href="static/css/style.css" type="text/css" />
     <link rel="stylesheet" href="https://unpkg.com/ionicons@4.5.10-0/dist/css/ionicons.min.css">
     <link rel="stylesheet" href="https://unpkg.com/mdui@1.0.2/dist/css/mdui.min.css" />
-    <link href="img/favicon.ico" rel="shortcut icon">
+    <link href="static/img/favicon.ico" rel="shortcut icon">
 </head>
 
 <body class="mdui-theme-primary-purple mdui-theme-accent-blue">
@@ -19,21 +19,8 @@
             <div id="content">
                 <div id="title-container">
                     <h1 id="page-title">Welcome to herta kuru</h1>
-                    <img id="title-img" alt="img of herta" src="img/favicon.ico">
+                    <img id="title-img" alt="img of herta" src="static/img/favicon.png">
                 </div>
-                <!-- <div id="language-selector-container">
-                        <select id="language-selector" class="mdui-select">
-                            <option value="en" selected>English</option>
-                            <option value="cn">中文</option>
-                            <option value="ja">日本語</option>
-                            <option value="kr">한국어</option>
-                            <option value="id">Bahasa Indonesia</option>
-                        </select>
-                        <select id="vo-language-selector" class="mdui-select">
-                            <option value="ja" selected>日本語</option>
-                            <option value="cn">中文</option>
-                        </select>
-                </div> -->
 
                 <h2 id="page-descriptions">
                     The website for Herta, the <del>annoying</del> cutest genius Honkai:
@@ -59,7 +46,7 @@
             </div>
         </main>
         <div id="footer">
-            <img id="herta-card" loading="lazy" src="img/card_en.jpg" class="img-autocdn" alt="" />
+            <img id="herta-card" loading="lazy" src="static/img/card_en.jpg" class="img-autocdn" alt="" />
             <div id="footer-text">
                 <div class="mdui-list"><a href="javascript:void(0);" class="mdui-list-item mdui-ripple"
                         id="show-credits-opt">
@@ -88,7 +75,7 @@
         </div>
     </div>
     <script src="https://unpkg.com/mdui@1.0.2/dist/js/mdui.min.js"></script>
-    <script src="script.js" async></script>
+    <script src="static/js/script.js" async></script>
 
     <!-- Google tag (gtag.js) -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-T303QPC9L9"></script>
